tcp-server/requirements.txt
redis[asyncio]==5.2.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.6.1
